📦 Chore : 헤더 컴포넌트 공동 컴포넌트 경로로 이동
